+++ content-cache/tests/integration/test_healthchecks.py
@@ -17,6 +17,8 @@ from tests.integration.helpers import (
     BACKENDS_PATH_CONFIG_NAME,
     HEALTHCHECK_INTERVAL_CONFIG_NAME,
     HEALTHCHECK_PATH_CONFIG_NAME,
+    HEALTHCHECK_SSL_VERIFY_CONFIG_NAME,
+    HEALTHCHECK_VALID_STATUS_CONFIG_NAME,
     HOSTNAME_CONFIG_NAME,
     PROTOCOL_CONFIG_NAME,
     PROXY_CACHE_VALID_CONFIG_NAME,
@@ -73,6 +75,8 @@ async def test_healthchecks_healthy(
     config[BACKENDS_PATH_CONFIG_NAME] = http_ok_path
     config[HEALTHCHECK_PATH_CONFIG_NAME] = "/health"
     config[HEALTHCHECK_INTERVAL_CONFIG_NAME] = str(HEALTHCHECK_INTERVAL)
+    config[HEALTHCHECK_SSL_VERIFY_CONFIG_NAME] = "false"
+    config[HEALTHCHECK_VALID_STATUS_CONFIG_NAME] = "200"
     config[PROTOCOL_CONFIG_NAME] = "http"
     config[PROXY_CACHE_VALID_CONFIG_NAME] = '["200 10s"]'
     await cache_tester.setup_config(config)